requirements.txt
# Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Supabase
supabase==2.3.4

# Anthropic Claude API
anthropic==0.18.1

# Environment variables
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
